♻️ [refactor]: #452 Phase 10: lexer_read_name.rs を分離
read_name の 39 テストを `src/lexer/tests/lexer_read_name.rs` に分離。
テスト本体で `PdfName::new` を直接参照するため
`use crate::object::name::PdfName;` を追加。

- 293 件グリーン維持
- 短縮テスト関数名集合不変 / 重複 0
- lexer_read_name.rs: 366 行 (< 800)

Refs #452
